test moved into src
