+++ typescript-id-allocator/src/test/copied-utils/stochastic/types.ts
