+++ typescript-id-allocator/src/test/id-compressor/idCompressor.spec.ts
@@ -6,6 +6,10 @@
 import { strict as assert } from "assert";
 import { MockLogger } from "@fluidframework/telemetry-utils";
 import { validateAssertionError } from "@fluidframework/test-runtime-utils";
+import { fail } from "../../../src/copied-utils";
+import { take } from "../copied-utils/stochastic";
+import { OpSpaceCompressedId, SessionSpaceCompressedId, StableId } from "../../../src/types";
+import { createSessionId } from "../../../src/utilities";
 import {
 	performFuzzActions,
 	sessionIds,
@@ -19,10 +23,6 @@ import {
 	CompressorFactory,
 } from "./idCompressorTestUtilities";
 import { compressorEquals, isFinalId, isLocalId } from "./testCommon";
-import { take } from "../copied-utils/stochastic";
-import { OpSpaceCompressedId, SessionSpaceCompressedId, StableId } from "../../src/types";
-import { createSessionId } from "../../src/utilities";
-import { fail } from "../../src/copied-utils";
 
 describe("IdCompressor", () => {
 	itCompressor("caches and evicts tokens (clearbox)", () => {
